refactor: Melhorar organização dos DTOs
