+++ backend/src/main/java/com/livraria/livraria/entities/dtos/response/BookDTO.java
@@ -1,0 +1,40 @@
+package com.livraria.livraria.entities.dtos.response;
+
+import java.io.Serializable;
+
+public class BookDTO implements Serializable{
+
+    private static final long serialVersionUID = 1L;
+
+    private Long id;
+    private String name;
+    private int amount;
+
+    public BookDTO() {
+    }
+
+    public BookDTO(Long id, String name) {
+        this.id = id;
+        this.name = name;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public int getAmount() {
+        return amount;
+    }
+
+    public void setAmount(int amount) {
+        this.amount = amount;
+    }
+}
